Move asymmetric algorithm modules into the toplevel
Makes them more discoverable, and the associated names shorter and more
convenient.
