+++ src/authentication/mod.rs
@@ -5,6 +5,6 @@ mod algorithm;
 pub mod commands;
 mod credentials;
 mod error;
-mod key;
+pub mod key;
 
-pub use self::{algorithm::Algorithm, credentials::*, error::*, key::*};
+pub use self::{algorithm::Algorithm, credentials::*, error::*, key::Key};
