+++ src/ed25519/commands.rs
