+++ src/ed25519/signer.rs
