+++ src/mockhsm/command.rs
@@ -3,23 +3,21 @@
 use super::{object::Payload, state::State, MOCK_SERIAL_NUMBER};
 use crate::{
     algorithm::*,
-    asymmetric::{
-        self,
-        commands::*,
-        ecdsa,
-        ed25519::{self, commands::*},
-        kex, rsa, PublicKey,
-    },
+    asymmetric::{self, commands::*, PublicKey},
     audit::{commands::*, AuditCommand, AuditOption, AuditTag},
     authentication::{self, commands::*},
     command::{Code, Message},
     connector::ConnectionError,
     device::{self, commands::*, DeviceErrorKind, SerialNumber, StorageInfo},
+    ecdsa,
+    ed25519::{self, commands::*},
     hmac::{self, commands::*},
+    kex,
     object::{self, commands::*},
     opaque::{self, commands::*},
     otp,
     response::{self, Response},
+    rsa,
     serialization::deserialize,
     session::{self, commands::*},
     template,
